# model: convert more *strings to strings