--- a/model/message_test.go
+++ b/model/message_test.go
@@ -37,8 +37,8 @@ func TestMessaging_Fields(t *testing.T) {
 	require.Equal(t, common.MapStr{}, m.Fields())
 
 	m = &Message{
-		QueueName: tests.StringPtr("orders"),
-		Body:      tests.StringPtr("order confirmed"),
+		QueueName: "orders",
+		Body:      "order confirmed",
 		Headers:   http.Header{"Internal": []string{"false"}, "Services": []string{"user", "order"}},
 		AgeMillis: tests.IntPtr(1577958057123),
 	}