--- a/tests/common.go
+++ b/tests/common.go
@@ -32,11 +32,6 @@ func Group(s string) group {
 	return group{str: s}
 }
 
-// StringPtr is a test helper function that returns the address of the given string
-func StringPtr(s string) *string {
-	return &s
-}
-
 // IntPtr is a test helper function that returns the address of the given integer
 func IntPtr(i int) *int {
 	return &i