… model/error_test.go
… model/stacktrace_frame_test.go
… model/stacktrace_test.go