## Motivation/summary

Simplify the model types by converting `*string` to `string` where we can rely on an empty string being invalid.

## How to test these changes

Non-functional change, run the tests.

## Related issues

None.