@axw 2021-02-23
I modified the processor/stream benchmark to also transform events, and compared with master. Benchstat results:

```
name                                                               old time/op    new time/op    delta
BackendProcessor/errors.ndjson/NoRateLimit-12                         172µs ± 1%     152µs ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/errors.ndjson/WithRateLimit-12                       178µs ± 6%     155µs ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/errors_2.ndjson/NoRateLimit-12                       127µs ± 1%     174µs ±31%   ~     (p=0.700 n=3+3)
BackendProcessor/errors_2.ndjson/WithRateLimit-12                     129µs ± 1%     135µs ±19%   ~     (p=0.700 n=3+3)
BackendProcessor/errors_rum.ndjson/NoRateLimit-12                    41.4µs ± 3%    36.8µs ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/errors_rum.ndjson/WithRateLimit-12                  49.6µs ±15%    39.6µs ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/events.ndjson/NoRateLimit-12                         307µs ± 1%     248µs ±11%   ~     (p=0.100 n=3+3)
BackendProcessor/events.ndjson/WithRateLimit-12                       314µs ± 1%     240µs ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/experimental.ndjson/NoRateLimit-12                  39.8µs ± 9%    35.9µs ± 6%   ~     (p=0.200 n=3+3)
BackendProcessor/experimental.ndjson/WithRateLimit-12                36.8µs ± 4%    39.1µs ± 7%   ~     (p=0.400 n=3+3)
BackendProcessor/heavy.ndjson/NoRateLimit-12                         22.4ms ±11%    20.8ms ± 5%   ~     (p=0.400 n=3+3)
BackendProcessor/heavy.ndjson/WithRateLimit-12                       24.2ms ±12%    19.8ms ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/NoRateLimit-12            7.82µs ± 2%    8.32µs ± 3%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/WithRateLimit-12          9.20µs ± 0%   10.04µs ± 7%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event.ndjson/NoRateLimit-12                 47.0µs ± 1%    51.0µs ± 5%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event.ndjson/WithRateLimit-12               49.8µs ± 2%    53.0µs ± 5%   ~     (p=0.200 n=3+3)
BackendProcessor/invalid-json-event.ndjson/NoRateLimit-12            20.4µs ± 1%    20.9µs ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-event.ndjson/WithRateLimit-12          22.8µs ± 2%    22.9µs ± 3%   ~     (p=1.000 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/NoRateLimit-12         20.3µs ± 1%    23.5µs ± 5%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/WithRateLimit-12       20.7µs ± 1%    22.8µs ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/NoRateLimit-12            3.90µs ± 2%    4.69µs ±11%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/WithRateLimit-12          3.96µs ± 1%    4.35µs ± 3%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata.ndjson/NoRateLimit-12              4.06µs ± 1%    4.69µs ± 4%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata.ndjson/WithRateLimit-12            4.07µs ± 1%    4.90µs ± 7%   ~     (p=0.100 n=3+3)
BackendProcessor/metadata-null-values.ndjson/NoRateLimit-12          14.9µs ± 4%    14.9µs ± 5%   ~     (p=1.000 n=3+3)
BackendProcessor/metadata-null-values.ndjson/WithRateLimit-12        16.1µs ± 1%    16.6µs ± 4%   ~     (p=0.400 n=3+3)
BackendProcessor/metadata.ndjson/NoRateLimit-12                      11.8µs ± 0%    12.7µs ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/metadata.ndjson/WithRateLimit-12                    13.4µs ± 2%    14.4µs ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/metricsets.ndjson/NoRateLimit-12                    63.2µs ± 1%    65.1µs ± 3%   ~     (p=0.200 n=3+3)
BackendProcessor/metricsets.ndjson/WithRateLimit-12                  63.1µs ± 5%    66.5µs ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/minimal-service.ndjson/NoRateLimit-12               20.9µs ± 6%    20.2µs ± 5%   ~     (p=0.700 n=3+3)
BackendProcessor/minimal-service.ndjson/WithRateLimit-12             20.8µs ± 4%    23.2µs ±13%   ~     (p=0.400 n=3+3)
BackendProcessor/minimal.ndjson/NoRateLimit-12                       55.1µs ± 9%    54.4µs ± 3%   ~     (p=1.000 n=3+3)
BackendProcessor/minimal.ndjson/WithRateLimit-12                     58.8µs ± 6%    53.4µs ± 4%   ~     (p=0.100 n=3+3)
BackendProcessor/optional-timestamps.ndjson/NoRateLimit-12           42.9µs ±12%    42.9µs ± 4%   ~     (p=0.700 n=3+3)
BackendProcessor/optional-timestamps.ndjson/WithRateLimit-12         48.4µs ± 7%    43.4µs ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/ratelimit.ndjson/NoRateLimit-12                      136µs ± 1%     146µs ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/ratelimit.ndjson/WithRateLimit-12                    139µs ± 1%     156µs ± 7%   ~     (p=0.100 n=3+3)
BackendProcessor/spans.ndjson/NoRateLimit-12                          150µs ± 3%     169µs ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/spans.ndjson/WithRateLimit-12                        151µs ± 1%     172µs ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions.ndjson/NoRateLimit-12                   164µs ± 1%     200µs ± 9%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions.ndjson/WithRateLimit-12                 167µs ± 4%     201µs ± 8%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans.ndjson/NoRateLimit-12             277µs ±19%     267µs ± 2%   ~     (p=0.700 n=3+3)
BackendProcessor/transactions_spans.ndjson/WithRateLimit-12           259µs ±19%     268µs ± 3%   ~     (p=0.700 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/NoRateLimit-12        40.9µs ± 2%    42.2µs ± 3%   ~     (p=0.200 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/WithRateLimit-12      43.7µs ± 8%    43.3µs ± 6%   ~     (p=1.000 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/NoRateLimit-12      40.1µs ± 1%    38.4µs ± 3%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/WithRateLimit-12    41.3µs ± 2%    43.1µs ± 6%   ~     (p=0.400 n=3+3)
RUMV3Processor/rum_errors.ndjson/NoRateLimit-12                      33.5µs ± 3%    33.0µs ± 2%   ~     (p=0.400 n=3+3)
RUMV3Processor/rum_errors.ndjson/WithRateLimit-12                    35.7µs ± 6%    35.0µs ± 5%   ~     (p=1.000 n=3+3)
RUMV3Processor/rum_events.ndjson/NoRateLimit-12                       237µs ± 4%     258µs ± 4%   ~     (p=0.100 n=3+3)
RUMV3Processor/rum_events.ndjson/WithRateLimit-12                     240µs ± 2%     239µs ± 2%   ~     (p=0.700 n=3+3)

name                                                               old speed      new speed      delta
BackendProcessor/errors.ndjson/NoRateLimit-12                      36.5MB/s ± 1%  41.3MB/s ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/errors.ndjson/WithRateLimit-12                    35.4MB/s ± 6%  40.4MB/s ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/errors_2.ndjson/NoRateLimit-12                    37.2MB/s ± 1%  28.8MB/s ±36%   ~     (p=0.700 n=3+3)
BackendProcessor/errors_2.ndjson/WithRateLimit-12                  36.6MB/s ± 1%  35.4MB/s ±18%   ~     (p=0.700 n=3+3)
BackendProcessor/errors_rum.ndjson/NoRateLimit-12                  45.9MB/s ± 3%  51.6MB/s ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/errors_rum.ndjson/WithRateLimit-12                38.7MB/s ±16%  48.0MB/s ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/events.ndjson/NoRateLimit-12                      24.2MB/s ± 1%  30.1MB/s ±11%   ~     (p=0.100 n=3+3)
BackendProcessor/events.ndjson/WithRateLimit-12                    23.7MB/s ± 1%  30.9MB/s ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/experimental.ndjson/NoRateLimit-12                21.3MB/s ±10%  23.5MB/s ± 6%   ~     (p=0.200 n=3+3)
BackendProcessor/experimental.ndjson/WithRateLimit-12              22.9MB/s ± 4%  21.6MB/s ± 7%   ~     (p=0.400 n=3+3)
BackendProcessor/heavy.ndjson/NoRateLimit-12                       17.9MB/s ±11%  19.2MB/s ± 5%   ~     (p=0.400 n=3+3)
BackendProcessor/heavy.ndjson/WithRateLimit-12                     16.7MB/s ±13%  20.2MB/s ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/NoRateLimit-12          50.0MB/s ± 2%  47.0MB/s ± 3%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/WithRateLimit-12        42.5MB/s ± 0%  39.0MB/s ± 6%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event.ndjson/NoRateLimit-12               16.3MB/s ± 1%  15.0MB/s ± 5%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event.ndjson/WithRateLimit-12             15.4MB/s ± 2%  14.5MB/s ± 5%   ~     (p=0.200 n=3+3)
BackendProcessor/invalid-json-event.ndjson/NoRateLimit-12          28.7MB/s ± 1%  28.1MB/s ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-event.ndjson/WithRateLimit-12        25.7MB/s ± 2%  25.7MB/s ± 3%   ~     (p=1.000 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/NoRateLimit-12       22.0MB/s ± 1%  19.1MB/s ± 6%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/WithRateLimit-12     21.6MB/s ± 1%  19.6MB/s ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/NoRateLimit-12           112MB/s ± 2%    93MB/s ±10%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/WithRateLimit-12         110MB/s ± 1%   100MB/s ± 3%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata.ndjson/NoRateLimit-12             110MB/s ± 1%    95MB/s ± 4%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata.ndjson/WithRateLimit-12           110MB/s ± 1%    91MB/s ± 7%   ~     (p=0.100 n=3+3)
BackendProcessor/metadata-null-values.ndjson/NoRateLimit-12        35.2MB/s ± 4%  35.3MB/s ± 5%   ~     (p=1.000 n=3+3)
BackendProcessor/metadata-null-values.ndjson/WithRateLimit-12      32.6MB/s ± 1%  31.6MB/s ± 4%   ~     (p=0.400 n=3+3)
BackendProcessor/metadata.ndjson/NoRateLimit-12                     103MB/s ± 0%    96MB/s ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/metadata.ndjson/WithRateLimit-12                  91.3MB/s ± 2%  84.5MB/s ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/metricsets.ndjson/NoRateLimit-12                  25.6MB/s ± 1%  24.8MB/s ± 3%   ~     (p=0.200 n=3+3)
BackendProcessor/metricsets.ndjson/WithRateLimit-12                25.7MB/s ± 6%  24.3MB/s ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/minimal-service.ndjson/NoRateLimit-12             20.4MB/s ± 6%  21.0MB/s ± 5%   ~     (p=0.700 n=3+3)
BackendProcessor/minimal-service.ndjson/WithRateLimit-12           20.4MB/s ± 4%  18.5MB/s ±13%   ~     (p=0.400 n=3+3)
BackendProcessor/minimal.ndjson/NoRateLimit-12                     18.8MB/s ± 8%  19.0MB/s ± 3%   ~     (p=1.000 n=3+3)
BackendProcessor/minimal.ndjson/WithRateLimit-12                   17.6MB/s ± 6%  19.3MB/s ± 4%   ~     (p=0.100 n=3+3)
BackendProcessor/optional-timestamps.ndjson/NoRateLimit-12         24.1MB/s ±12%  24.0MB/s ± 4%   ~     (p=0.700 n=3+3)
BackendProcessor/optional-timestamps.ndjson/WithRateLimit-12       21.3MB/s ± 7%  23.7MB/s ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/ratelimit.ndjson/NoRateLimit-12                   30.9MB/s ± 1%  28.8MB/s ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/ratelimit.ndjson/WithRateLimit-12                 30.3MB/s ± 1%  27.0MB/s ± 7%   ~     (p=0.100 n=3+3)
BackendProcessor/spans.ndjson/NoRateLimit-12                       28.3MB/s ± 3%  25.2MB/s ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/spans.ndjson/WithRateLimit-12                     28.2MB/s ± 1%  24.7MB/s ± 2%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions.ndjson/NoRateLimit-12                28.6MB/s ± 1%  23.6MB/s ± 9%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions.ndjson/WithRateLimit-12              28.2MB/s ± 4%  23.5MB/s ± 8%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans.ndjson/NoRateLimit-12          21.3MB/s ±17%  21.8MB/s ± 2%   ~     (p=0.700 n=3+3)
BackendProcessor/transactions_spans.ndjson/WithRateLimit-12        22.9MB/s ±18%  21.7MB/s ± 3%   ~     (p=0.700 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/NoRateLimit-12      27.3MB/s ± 2%  26.5MB/s ± 3%   ~     (p=0.200 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/WithRateLimit-12    25.6MB/s ± 8%  25.8MB/s ± 5%   ~     (p=1.000 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/NoRateLimit-12    27.4MB/s ± 1%  28.6MB/s ± 3%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/WithRateLimit-12  26.5MB/s ± 2%  25.5MB/s ± 6%   ~     (p=0.400 n=3+3)
RUMV3Processor/rum_errors.ndjson/NoRateLimit-12                    28.6MB/s ± 3%  29.1MB/s ± 2%   ~     (p=0.400 n=3+3)
RUMV3Processor/rum_errors.ndjson/WithRateLimit-12                  26.9MB/s ± 6%  27.4MB/s ± 5%   ~     (p=1.000 n=3+3)
RUMV3Processor/rum_events.ndjson/NoRateLimit-12                    17.0MB/s ± 4%  15.6MB/s ± 4%   ~     (p=0.100 n=3+3)
RUMV3Processor/rum_events.ndjson/WithRateLimit-12                  16.7MB/s ± 2%  16.8MB/s ± 2%   ~     (p=0.700 n=3+3)

name                                                               old alloc/op   new alloc/op   delta
BackendProcessor/errors.ndjson/NoRateLimit-12                         118kB ± 1%     111kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/errors.ndjson/WithRateLimit-12                       118kB ± 0%     113kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/errors_2.ndjson/NoRateLimit-12                      85.8kB ± 0%    80.4kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/errors_2.ndjson/WithRateLimit-12                    86.2kB ± 0%    80.1kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/errors_rum.ndjson/NoRateLimit-12                    24.4kB ± 0%    20.9kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/errors_rum.ndjson/WithRateLimit-12                  24.5kB ± 1%    21.4kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/events.ndjson/NoRateLimit-12                         142kB ± 1%     139kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/events.ndjson/WithRateLimit-12                       144kB ± 0%     139kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/experimental.ndjson/NoRateLimit-12                  24.0kB ± 0%    24.0kB ± 1%   ~     (p=1.000 n=3+3)
BackendProcessor/experimental.ndjson/WithRateLimit-12                24.4kB ± 1%    24.4kB ± 1%   ~     (p=1.000 n=3+3)
BackendProcessor/heavy.ndjson/NoRateLimit-12                         18.9MB ± 0%    18.7MB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/heavy.ndjson/WithRateLimit-12                       18.9MB ± 0%    18.8MB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/NoRateLimit-12            2.50kB ± 0%    2.48kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/WithRateLimit-12          2.82kB ± 2%    2.82kB ± 1%   ~     (p=0.700 n=3+3)
BackendProcessor/invalid-event.ndjson/NoRateLimit-12                 16.5kB ± 0%    16.8kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event.ndjson/WithRateLimit-12               17.0kB ± 1%    17.2kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-event.ndjson/NoRateLimit-12            11.4kB ± 1%    11.4kB ± 1%   ~     (p=0.400 n=3+3)
BackendProcessor/invalid-json-event.ndjson/WithRateLimit-12          11.7kB ± 0%    11.8kB ± 1%   ~     (p=0.200 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/NoRateLimit-12         4.63kB ± 1%    4.66kB ± 1%   ~     (p=0.700 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/WithRateLimit-12       4.65kB ± 1%    4.69kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/NoRateLimit-12            1.61kB ± 0%    1.62kB ± 1%   ~     (p=0.700 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/WithRateLimit-12          1.62kB ± 1%    1.63kB ± 1%   ~     (p=0.400 n=3+3)
BackendProcessor/invalid-metadata.ndjson/NoRateLimit-12              1.61kB ± 1%    1.62kB ± 1%   ~     (p=0.300 n=3+3)
BackendProcessor/invalid-metadata.ndjson/WithRateLimit-12            1.62kB ± 0%    1.63kB ± 1%   ~     (p=0.200 n=3+3)
BackendProcessor/metadata-null-values.ndjson/NoRateLimit-12          7.54kB ± 0%    7.18kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/metadata-null-values.ndjson/WithRateLimit-12        7.86kB ± 1%    7.55kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/metadata.ndjson/NoRateLimit-12                      3.14kB ± 1%    3.13kB ± 0%   ~     (p=0.400 n=3+3)
BackendProcessor/metadata.ndjson/WithRateLimit-12                    3.47kB ± 1%    3.49kB ± 2%   ~     (p=0.700 n=3+3)
BackendProcessor/metricsets.ndjson/NoRateLimit-12                    33.6kB ± 1%    34.2kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/metricsets.ndjson/WithRateLimit-12                  33.8kB ± 1%    34.7kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/minimal-service.ndjson/NoRateLimit-12               11.8kB ± 0%    11.4kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/minimal-service.ndjson/WithRateLimit-12             12.1kB ± 1%    11.8kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/minimal.ndjson/NoRateLimit-12                       36.3kB ± 1%    36.7kB ± 0%   ~     (p=0.200 n=3+3)
BackendProcessor/minimal.ndjson/WithRateLimit-12                     36.6kB ± 0%    37.0kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/optional-timestamps.ndjson/NoRateLimit-12           29.6kB ± 0%    30.0kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/optional-timestamps.ndjson/WithRateLimit-12         29.8kB ± 0%    30.4kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/ratelimit.ndjson/NoRateLimit-12                      125kB ± 0%     125kB ± 0%   ~     (p=1.000 n=3+3)
BackendProcessor/ratelimit.ndjson/WithRateLimit-12                    125kB ± 0%     125kB ± 0%   ~     (p=1.000 n=3+3)
BackendProcessor/spans.ndjson/NoRateLimit-12                          130kB ± 0%     131kB ± 0%   ~     (p=0.200 n=3+3)
BackendProcessor/spans.ndjson/WithRateLimit-12                        131kB ± 0%     131kB ± 0%   ~     (p=0.700 n=3+3)
BackendProcessor/transactions.ndjson/NoRateLimit-12                   104kB ± 0%     105kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions.ndjson/WithRateLimit-12                 105kB ± 0%     107kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans.ndjson/NoRateLimit-12             171kB ± 0%     173kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans.ndjson/WithRateLimit-12           171kB ± 0%     174kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/NoRateLimit-12        25.3kB ± 0%    24.6kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/WithRateLimit-12      25.7kB ± 1%    24.9kB ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/NoRateLimit-12      26.5kB ± 1%    25.4kB ± 1%   ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/WithRateLimit-12    26.8kB ± 1%    25.8kB ± 1%   ~     (p=0.100 n=3+3)
RUMV3Processor/rum_errors.ndjson/NoRateLimit-12                      18.5kB ± 0%    16.7kB ± 0%   ~     (p=0.100 n=3+3)
RUMV3Processor/rum_errors.ndjson/WithRateLimit-12                    18.8kB ± 0%    17.0kB ± 0%   ~     (p=0.100 n=3+3)
RUMV3Processor/rum_events.ndjson/NoRateLimit-12                       177kB ± 0%     176kB ± 0%   ~     (p=0.400 n=3+3)
RUMV3Processor/rum_events.ndjson/WithRateLimit-12                     177kB ± 0%     176kB ± 0%   ~     (p=0.200 n=3+3)

name                                                               old allocs/op  new allocs/op  delta
BackendProcessor/errors.ndjson/NoRateLimit-12                         1.34k ± 0%     1.25k ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/errors.ndjson/WithRateLimit-12                       1.35k ± 0%     1.25k ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/errors_2.ndjson/NoRateLimit-12                       1.00k ± 0%     0.93k ± 0%   ~     (p=0.100 n=3+3)
BackendProcessor/errors_2.ndjson/WithRateLimit-12                     1.01k ± 0%     0.94k ± 0%   ~     (p=0.100 n=3+3)
…